restruct new router.

+++ dialer/balance.go
@@ -191,7 +191,7 @@ func (b *BalancedDialer) Matched(uri string) bool {
 	return b.matcher.MatchString(uri)
 }
 
-func (b *BalancedDialer) Dial(channel Channel, sid uint64, uri string, pipe io.ReadWriteCloser) (r Conn, err error) {
+func (b *BalancedDialer) Dial(channel Channel, sid uint16, uri string, pipe io.ReadWriteCloser) (r Conn, err error) {
 	for _, f := range b.Filters {
 		if f.Matcher.MatchString(uri) {
 			if f.Access < 1 {
