+++ dialer/tcp.go
@@ -46,7 +46,7 @@ func (t *TCPDialer) Matched(uri string) bool {
 }
 
 // Dial one connection by uri
-func (t *TCPDialer) Dial(channel Channel, sid uint64, uri string, pipe io.ReadWriteCloser) (raw Conn, err error) {
+func (t *TCPDialer) Dial(channel Channel, sid uint16, uri string, pipe io.ReadWriteCloser) (raw Conn, err error) {
 	remote, err := url.Parse(uri)
 	if err == nil {
 		var dialer net.Dialer
